# Improved logging