--- a/build_runner/lib/src/entrypoint/serve.dart
+++ b/build_runner/lib/src/entrypoint/serve.dart
@@ -13,7 +13,6 @@ import 'package:shelf/shelf_io.dart';
 
 import '../build_script_generate/build_process_state.dart';
 import '../generate/build.dart';
-import '../server/server.dart';
 import 'options.dart';
 import 'watch.dart';