… _test/test/build_integration_test.dart
… _test/test/common/utils.dart
… _test/test/goldens/generated_build_script.dart
… _test_common/lib/test_environment.dart
… _test_common/lib/test_phases.dart
… build/CHANGELOG.md
… build_modules/pubspec.yaml
… build_modules/test/build_test.dart
… build_modules/test/meta_module_clean_builder_test.dart
… build_resolvers/CHANGELOG.md
… build_resolvers/lib/src/human_readable_duration.dart
… build_runner/CHANGELOG.md
… build_runner/lib/src/build_script_generate/build_process_state.dart
… build_runner/lib/src/build_script_generate/build_script_generate.dart
… build_runner/lib/src/entrypoint/base_command.dart
… build_runner/lib/src/entrypoint/build.dart
… build_runner/lib/src/entrypoint/clean.dart
… build_runner/lib/src/entrypoint/daemon.dart
… build_runner/lib/src/entrypoint/doctor.dart
… build_runner/lib/src/entrypoint/options.dart
… build_runner/lib/src/entrypoint/run_script.dart
… build_runner/lib/src/entrypoint/test.dart
… build_runner/lib/src/entrypoint/watch.dart
… build_runner/lib/src/generate/build.dart
… build_runner/lib/src/generate/watch_impl.dart
… build_runner/lib/src/internal.dart
… build_runner/lib/src/logging/std_io_logging.dart
… build_runner/lib/src/package_graph/build_config_overrides.dart
… build_runner/lib/src/server/server.dart
… build_runner/lib/src/watcher/graph_watcher.dart