For #4005.

This is a rewrite of the logging to provide more information about what's happening and particularly about where the time is going.

High level changes:

 - Two output modes: "console" mode that used ANSI codes to rewrite lines and format text, and "line at a time".
 - The output now shows the status of each builder, and where the time is being spent for that builder. Timings, which appear per builder as e.g. "spent 4s analyzing", are only shown for timings >=1s, which should be plenty for diagnosing slow builds, we can make this configurable if needed.
 - The only thing the `verbose` command line flag does now is to change what is logged from user code (builders). By default only WARNING and higher is logged from user code, if `verbose` then everything is logged.
 - Removed the disabled-by-default "why this build step ran" logging; it's less important and can be added back later on top of the new logging if needed.

## Before/after examples

### json_serializable, no console

before

```
[INFO] Generating build script...
[INFO] Generating build script completed, took 342ms

[INFO] Precompiling build script......
[INFO] Precompiling build script... completed, took 5.8s

[INFO] Initializing inputs
[INFO] Building new asset graph...
[INFO] Building new asset graph completed, took 792ms

[INFO] Checking for unexpected pre-existing outputs....
[INFO] Checking for unexpected pre-existing outputs. completed, took 1ms

[INFO] Running build...
[INFO] Generating SDK summary...
[INFO] 2.3s elapsed, 0/16 actions completed.
[INFO] 4.3s elapsed, 0/16 actions completed.
[INFO] Generating SDK summary completed, took 5.0s

[WARNING] json_serializable on lib/lib500.dart:
The version constraint "any" on json_annotation allows versions before 4.9.0 which is not allowed.
[INFO] 6.5s elapsed, 112/126 actions completed.
[INFO] 7.5s elapsed, 198/214 actions completed.
[INFO] 8.6s elapsed, 274/290 actions completed.
[INFO] 9.6s elapsed, 334/348 actions completed.
[INFO] 10.7s elapsed, 395/409 actions completed.
[INFO] 11.8s elapsed, 450/466 actions completed.
[INFO] 12.9s elapsed, 504/519 actions completed.
[INFO] 13.9s elapsed, 691/707 actions completed.
[INFO] 15.0s elapsed, 810/815 actions completed.
[INFO] 16.1s elapsed, 895/905 actions completed.
[INFO] 17.1s elapsed, 963/965 actions completed.
[INFO] Running build completed, took 17.8s

[INFO] Caching finalized dependency graph...
[INFO] Caching finalized dependency graph completed, took 281ms

[INFO] Succeeded after 18.2s with 1002 outputs (1506 actions)
```

before, incremental

```
[INFO] Generating build script...
[INFO] Generating build script completed, took 324ms

[INFO] Initializing inputs
[INFO] Reading cached asset graph...
[INFO] Reading cached asset graph completed, took 657ms

[INFO] Checking for updates since last build...
[INFO] Checking for updates since last build completed, took 1.4s

[INFO] Running build...
[INFO] 1.1s elapsed, 9/25 actions completed.
[WARNING] json_serializable on lib/lib494.dart:
The version constraint "any" on json_annotation allows versions before 4.9.0 which is not allowed.
[INFO] 2.1s elapsed, 94/110 actions completed.
[INFO] 3.1s elapsed, 174/189 actions completed.
[INFO] 4.1s elapsed, 244/259 actions completed.
[INFO] 5.2s elapsed, 305/320 actions completed.
[INFO] 6.2s elapsed, 361/375 actions completed.
[INFO] 7.3s elapsed, 413/427 actions completed.
[INFO] 8.4s elapsed, 458/473 actions completed.
[INFO] 9.4s elapsed, 503/510 actions completed.
[INFO] 10.6s elapsed, 681/695 actions completed.
[INFO] 11.7s elapsed, 783/791 actions completed.
[INFO] 12.7s elapsed, 851/867 actions completed.
[INFO] 13.8s elapsed, 919/920 actions completed.
[INFO] 14.8s elapsed, 970/986 actions completed.
[INFO] Running build completed, took 15.4s

[INFO] Caching finalized dependency graph...
[INFO] Caching finalized dependency graph completed, took 209ms

[INFO] Succeeded after 15.8s with 1009 outputs (1019 actions)
```

after

```
  Generating the build script.
  Compiling the build script.
  Reading the asset graph.
  Creating the asset graph.
  Doing initial build cleanup.
  Updating the asset graph.
  Building, full build because builders changed.
  0s json_serializable on 501 inputs; lib/app.dart
  5s json_serializable on 501 inputs: 1 no-op; spent 5s sdk; lib/lib001.dart
W json_serializable on lib/lib001.dart:
  The version constraint "any" on json_annotation allows versions before 4.9.0 which is not allowed.
  7s json_serializable on 501 inputs: 1 output, 1 no-op; spent 5s sdk, 1s analyzing; lib/lib002.dart
  8s json_serializable on 501 inputs: 194 output, 1 no-op; spent 5s sdk, 2s analyzing; lib/lib195.dart
  9s json_serializable on 501 inputs: 425 output, 1 no-op; spent 5s sdk, 3s analyzing; lib/lib426.dart
  9s json_serializable on 501 inputs: 500 output, 1 no-op; spent 5s sdk, 3s analyzing
  0s source_gen:combining_builder on 501 inputs; lib/app.dart
  1s source_gen:combining_builder on 501 inputs: 169 output, 1 no-op; lib/lib170.dart
  2s source_gen:combining_builder on 501 inputs: 454 output, 1 no-op; spent 1s building; lib/lib455.dart
  2s source_gen:combining_builder on 501 inputs: 500 output, 1 no-op; spent 1s building
  Running the post build.
  Writing the asset graph.
  Built with build_runner in 13s; wrote 1002 outputs.
```

after, incremental

```
  Generating the build script.
  Reading the asset graph.
  Checking for updates.
  Updating the asset graph.
  Building, incremental build.
  0s json_serializable on 501 inputs; lib/app.dart
  0s json_serializable on 501 inputs: 1 no-op; lib/lib001.dart
W json_serializable on lib/lib001.dart:
  The version constraint "any" on json_annotation allows versions before 4.9.0 which is not allowed.
  3s json_serializable on 501 inputs: 1 same, 1 no-op; spent 2s analyzing; lib/lib002.dart
  4s json_serializable on 501 inputs: 147 same, 1 no-op; spent 3s analyzing; lib/lib148.dart
  5s json_serializable on 501 inputs: 348 same, 1 no-op; spent 3s analyzing; lib/lib349.dart
  5s json_serializable on 501 inputs: 500 same, 1 no-op; spent 4s analyzing
  0s source_gen:combining_builder on 501 inputs; lib/app.dart
  0s source_gen:combining_builder on 501 inputs: 1 skipped, 69 same; lib/lib070.dart
  1s source_gen:combining_builder on 501 inputs: 1 skipped, 314 same; lib/lib315.dart
  2s source_gen:combining_builder on 501 inputs: 1 skipped, 500 same
  Running the post build.
  Writing the asset graph.
  Built with build_runner in 9s; wrote 1009 outputs.
```

### json_serializable, console

before

```
[INFO] Generating build script completed, took 360ms
[INFO] Precompiling build script... completed, took 5.2s
[INFO] Building new asset graph completed, took 864ms
[INFO] Checking for unexpected pre-existing outputs. completed, took 1ms
[INFO] Generating SDK summary completed, took 4.9s
[WARNING] json_serializable on lib/lib499.dart:
The version constraint "any" on json_annotation allows versions before 4.9.0 which is not allowed.
[INFO] Running build completed, took 18.2s
[INFO] Caching finalized dependency graph completed, took 244ms
[INFO] Succeeded after 18.6s with 1002 outputs (1506 actions)
```

before, incremental

```
[INFO] Generating build script completed, took 335ms
[INFO] Reading cached asset graph completed, took 885ms
[INFO] Checking for updates since last build completed, took 1.4s
[WARNING] json_serializable on lib/lib499.dart:
The version constraint "any" on json_annotation allows versions before 4.9.0 which is not allowed.
[INFO] Running build completed, took 17.7s
[INFO] Caching finalized dependency graph completed, took 218ms
[INFO] Succeeded after 18.0s with 1009 outputs (1019 actions)
```

after

```
9s json_serializable on 501 inputs: 500 output, 1 no-op; spent 5s sdk, 3s analyzing                                                                                                                                                                          
2s source_gen:combining_builder on 501 inputs: 500 output, 1 no-op; spent 1s building                                                                                                                                                                        
                                                                                                                                                                                                                                                             
Built with build_runner in 13s with warnings; wrote 1002 outputs.                                                                                                                                                                                                                
                                                                                                                                                                                                                                                             
log output for json_serializable on lib/lib001.dart                                                                                                                                                                                                          
W The version constraint "any" on json_annotation allows versions before 4.9.0 which is not allowed. 
```

after, incremental

```
5s json_serializable on 501 inputs: 500 same, 1 no-op; spent 4s analyzing                                                                                                                                                                                    
2s source_gen:combining_builder on 501 inputs: 1 skipped, 500 same                                                                                                                                                                                           
                                                                                                                                                                                                                                                             
Built with build_runner in 9s with warnings; wrote 1009 outputs.                                                                                                                                                                                                                 
                                                                                                                                                                                                                                                             
log output for json_serializable on lib/lib001.dart                                                                                                                                                                                                          
W The version constraint "any" on json_annotation allows versions before 4.9.0 which is not allowed.
```

### web, no console

before

```
[INFO] Generating build script...
[INFO] Generating build script completed, took 347ms

[INFO] Precompiling build script......
[INFO] Precompiling build script... completed, took 6.2s

[INFO] Initializing inputs
[INFO] Building new asset graph...
[INFO] Building new asset graph completed, took 1.6s

[INFO] Checking for unexpected pre-existing outputs....
[INFO] Checking for unexpected pre-existing outputs. completed, took 1ms

[INFO] Running build...
[INFO] 1.1s elapsed, 58/60 actions completed.
[INFO] 2.2s elapsed, 58/60 actions completed.
[INFO] 3.3s elapsed, 58/60 actions completed.
[INFO] 7.6s elapsed, 2273/2275 actions completed.
[INFO] 8.6s elapsed, 2290/2298 actions completed.
[INFO] 9.7s elapsed, 2296/2298 actions completed.
[INFO] 10.8s elapsed, 2445/2450 actions completed.
[INFO] Running build completed, took 11.4s

[INFO] Caching finalized dependency graph...
[INFO] Caching finalized dependency graph completed, took 492ms

[INFO] Succeeded after 12.5s with 2521 outputs (7107 actions)
```

before, incremental

```
[INFO] Generating build script...
[INFO] Generating build script completed, took 348ms

[INFO] Initializing inputs
[INFO] Reading cached asset graph...
[INFO] Reading cached asset graph completed, took 717ms

[INFO] Checking for updates since last build...
[INFO] Checking for updates since last build completed, took 652ms

[INFO] Updating asset graph...
[INFO] Updating asset graph completed, took 8ms

[INFO] Running build...
[INFO] Running build completed, took 1.1s

[INFO] Caching finalized dependency graph...
[INFO] Caching finalized dependency graph completed, took 504ms

[INFO] Succeeded after 1.7s with 31 outputs (17 actions)
```

after

```
  Generating the build script.
  Compiling the build script.
  Reading the asset graph.
  Creating the asset graph.
  Doing initial build cleanup.
  Updating the asset graph.
  Building, full build because builders changed.
  0s provides_builder:some_builder on 27 inputs; lib/app.dart
  0s provides_builder:some_builder on 27 inputs: 1 output; test/build_integration_test.dart
  0s provides_builder:some_builder on 27 inputs: 27 output
  0s build_test:test_bootstrap on 61 inputs; $package$
  0s build_test:test_bootstrap on 61 inputs: 61 output
  0s build_web_compilers:entrypoint on 36 inputs; test/configurable_uri_test.dart.browser_test.dart
  3s build_web_compilers:sdk_js (lazy): 1 output; spent 3s building; build_web_compilers|$package$
  1s build_modules:module_library (lazy): 28 skipped, 473 output; analyzer|lib/src/fasta/error_converter.dart
  0s build_web_compilers:ddc_modules (lazy): 22 output; package_config|lib/$lib$
  3s build_modules:module_library (lazy): 28 skipped, 1417 output, 19 no-op; spent 2s building; web|lib/src/dom/webcodecs_vp9_codec_registration.dart
  4s build_modules:module_library (lazy): 28 skipped, 1991 output, 122 no-op; spent 3s building; csslib|lib/src/tree.dart
  1s build_web_compilers:ddc (lazy): 71 output; matcher|lib/src/core_matchers.ddc.module
  0s build_web_compilers:entrypoint on 36 inputs: 5 output; test/hello_world_custom_html_test.dart.browser_test.dart
  0s build_web_compilers:entrypoint on 36 inputs: 36 output
  Running the post build.
  Writing the asset graph.
  Built with build_runner in 15s; wrote 2521 outputs.
```

after, incremental

```
  Generating the build script.
  Reading the asset graph.
  Checking for updates.
  Updating the asset graph.
  Building, incremental build.
  0s provides_builder:some_builder on 27 inputs; lib/app.dart
  0s provides_builder:some_builder on 27 inputs: 1 output; test/build_integration_test.dart
  0s provides_builder:some_builder on 27 inputs: 26 skipped, 1 output
  0s build_test:test_bootstrap on 61 inputs; $package$
  0s build_test:test_bootstrap on 61 inputs: 61 skipped
  0s build_web_compilers:entrypoint on 36 inputs; test/configurable_uri_test.dart.browser_test.dart
  0s build_web_compilers:entrypoint on 36 inputs: 8 skipped, 1 output; test/hello_world_deferred_test.dart.browser_test.dart
  0s build_web_compilers:entrypoint on 36 inputs: 16 skipped, 20 output
  Running the post build.
  Writing the asset graph.
  Built with build_runner in 4s; wrote 31 outputs.
```

### web, console

before

```
[INFO] Generating build script completed, took 356ms
[INFO] Precompiling build script... completed, took 6.6s
[INFO] Building new asset graph completed, took 1.6s
[INFO] Checking for unexpected pre-existing outputs. completed, took 0ms
[INFO] Running build completed, took 10.8s
[INFO] Caching finalized dependency graph completed, took 418ms
[INFO] Succeeded after 11.8s with 2521 outputs (7107 actions)
```

before, incremental

```
[INFO] Generating build script completed, took 346ms
[INFO] Reading cached asset graph completed, took 611ms
[INFO] Checking for updates since last build completed, took 684ms
[INFO] Updating asset graph completed, took 8ms
[INFO] Running build completed, took 1.0s
[INFO] Caching finalized dependency graph completed, took 483ms
[INFO] Succeeded after 1.5s with 11 outputs (11 actions)
```

after

```
0s provides_builder:some_builder on 27 inputs: 27 output                                                                                                                                                                                                     
0s build_test:test_bootstrap on 61 inputs: 61 output                                                                                                                                                                                                         
0s build_web_compilers:entrypoint on 36 inputs: 36 output                                                                                                                                                                                                    
3s build_web_compilers:sdk_js (lazy): 2 output; spent 3s building                                                                                                                                                                                            
0s build_web_compilers:ddc_modules (lazy): 162 output                                                                                                                                                                                                        
4s build_modules:module_library (lazy): 28 skipped, 2011 output, 123 no-op; spent 3s building                                                                                                                                                                
4s build_web_compilers:ddc (lazy): 120 output; spent 4s building                                                                                                                                                                                             
                                                                                                                                                                                                                                                             
Built with build_runner in 15s; wrote 2521 outputs.
```

after, incremental

```
0s provides_builder:some_builder on 27 inputs: 26 skipped, 1 output                                                                                                                                                                                          
0s build_test:test_bootstrap on 61 inputs: 61 skipped                                                                                                                                                                                                        
0s build_web_compilers:entrypoint on 36 inputs: 36 skipped                                                                                                                                                                                                   
0s build_web_compilers:ddc_modules (lazy): 162 skipped                                                                                                                                                                                                       
0s build_modules:module_library (lazy): 2160 skipped, 2 same                                                                                                                                                                                                 
…